Maintain a database of installed projects
Backup, Generate, Read database (healthy, fail)
Each child process logs the state of project.
Latest state is written latest, and overrides all previous data.
A deleted project has value ``None``.

+++ pspman/define.py
@@ -238,7 +238,8 @@ def lock(env: InstallEnv, unlock: bool = False):
             for filetype in "healthy", "fail":
                 backup_file = os.path.join(env.clone_dir,
                                            f".pspman.{filetype}.yml")
-                if os.path.isfile(backup_file + ".bak"):
+                if os.path.isfile(backup_file + ".bak") and \
+                   not os.path.isfile(backup_file):
                     os.rename(backup_file + ".bak", backup_file)
             temp_build = os.path.join(env.prefix, 'temp_build')
             if os.path.isdir(temp_build):
